requirements.txt
streamlit
streamlit-option-menu
pandas
google-auth-oauthlib
google-api-python-client
